test: Move create channel split test into direct funding test
